functions/pipeline/requirements.txt
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
azure-storage-blob==1.4.0
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
grpcio==1.14.2
grpcio-tools==1.14.2
protobuf==3.6.1
six==1.11.0
pg8000==1.12.3
Pillow==5.3.0
urlpath==1.1.4
